refactor(direct-control): extract operation stable json helper
Refs: openspec/changes/civ7-support-direct-control-modularization

Move the duplicated stable JSON comparison helper into `src/play/operations/stable-json.ts` and reuse it from the unit, population, production, diplomacy, narrative, and unit-target postcondition paths. The helper preserves the existing array-aware key flattening behavior used for validation-drift, probe, and unit-target post-send comparisons.

This is local package/source relocation cleanup only; no runtime/live-game proof is claimed and Task 2.9.4 remains open. This does not change classifier ordering, validation-drift semantics, probe comparison semantics, unit-target no-repeat-after-unverified behavior, postcondition classification shape, telemetry, AI ingestion, CLI semantic projection, hotseat runtime proof, schema migration, debug hierarchy implementation, or Effect/oRPC procedure-core work.

Validation:
- bun run --cwd packages/civ7-direct-control test -- unit-operation.test.ts population-placement.test.ts production-choice.test.ts diplomacy-response.test.ts narrative-choice.test.ts unit-target-action.test.ts
- bun run --cwd packages/cli test -- game.play.operation-wrappers.test.ts game.play.production.test.ts game.play.diplomacy-response.test.ts game.play.narrative.test.ts game.play.unit-target.test.ts game/play/priorities.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run openspec:validate
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- git diff --check
- ownership/no-cycle scans for operation stable JSON helper ownership and Task 2.9.4 pending status
