+++ packages/civ7-direct-control/src/play/operations/diplomacy-postconditions.ts
@@ -1,3 +1,4 @@
+import { isRecord, stableJson } from "./stable-json.js";
 import type { Civ7DiplomacyResponseInput } from "./diplomacy-request.js";
 import type { Civ7OperationValidationResult } from "./types.js";
 import type { Civ7ComponentId } from "../../civ7-component-id.js";
@@ -118,26 +119,6 @@ function sameComponentId(left: Civ7ComponentId | null | undefined, right: Civ7Co
 
 function probeValue<T>(probe: Civ7RuntimeProbe<T>): T | undefined {
   return probe.ok ? probe.value : undefined;
-}
-
-function stableJson(value: unknown): string {
-  return JSON.stringify(value, Object.keys(flattenKeys(value)).sort()) ?? String(value);
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-
-function flattenKeys(value: unknown, keys: Record<string, true> = {}): Record<string, true> {
-  if (Array.isArray(value)) {
-    for (const item of value) flattenKeys(item, keys);
-  } else if (isRecord(value)) {
-    for (const [key, child] of Object.entries(value)) {
-      keys[key] = true;
-      flattenKeys(child, keys);
-    }
-  }
-  return keys;
 }
 
 async function sleep(ms: number): Promise<void> {
